+++ lib/authrocket/login_policy.rb
@@ -7,7 +7,7 @@ module AuthRocket
 
     attr :custom_domains, :external_css
     attr :footer, :header, :login_handler, :name, :primary_domain
-    attr :signup_handler, :subdomain
+    attr :redirect_uris, :signup_handler, :subdomain
     attr :base_domain, :domains # readonly
 
   end