+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 #### 2.0.0.pre2
 - Add AppHook#email_from_name
 - Add AuthProvider#authorize_token
+- Detect new ko_ API keys.
 
 #### 2.0.0.pre1
 - NOTE: This version includes breaking changes.